Merge remove_paste.py and and zerobin.py in cmd.py
The two commands have been merged together, using subcommands.

`remove_paste` is now accessible with `zerobin delete-paste`

`zerobin` is moved to `zerobin runserver` but is still accessible
with `zerobin` to keep compatibility with older versions. Unfortunately,
clize is not structured to allow default subcommand and the actual
implementation is nothing but a hack.